feat: Add unit- and integration tests. / refactor

+++ migrations/sqls/20220721180753-order-items-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE order_items (
     id SERIAL PRIMARY KEY,
-    order_id BIGINT REFERENCES orders(id),
-    product_id BIGINT REFERENCES products(id),
-    quantity INTEGER NOT NULL
+    order_id INT REFERENCES orders(id),
+    product_id INT REFERENCES products(id),
+    quantity INT NOT NULL
 );
